New desktop header (#4598)
### Description

This PR is an initial version of the desktop menu for the homepage
revamp. The mobile menu is temporarily removed, so that the desktop menu
can be reviewed in isolation.

*This PR is not intended for merge into master until the completion of
the homepage revamp*

### Notes

There is still some refactoring and cleanup to do (particularly in the
CSS) that will be done as part of a subsequent larger PR that includes
the mobile menu. However, the functionality of the desktop can be
previewed.

---------

Co-authored-by: Nicholas Blumberg <41446765+nick-next@users.noreply.github.com>
Co-authored-by: Nick B <nicholas.blumberg@gmail.com>
Co-authored-by: Pablo Noel <pablo@next-solutions.ca>